urbanpiper/requirements.txt
Django==3.1.7
uritemplate==3.0.1
wheel
djangorestframework==3.12.4
drf-yasg==1.20.0
apscheduler==3.7.0
model_bakery==1.2.1
sphinx==3.5.3